deleted the treecell versions of the DataClasses as its all now handled in the ItemShellTreeCell. Added a field for the proglangs in the CHIData (a reply Class) filled out the NewFXSwingMain GUI a bit more changed the duedate field in codehandin to a long updated the duedate method in MoodleClient that converts the duedate number into a conventional date form added separate methods for reading IOE file contents and added methods for getting the IOE files

+++ src/moodleclient/DataClasses/Codehandin.java
@@ -26,7 +26,7 @@ public class Codehandin {
     private int contextid;
     private String assignname;
     private String intro;
-    private int duedate;
+    private long duedate;
     private int funcpercent;
     private boolean spectestonly;
     /*
@@ -114,7 +114,7 @@ public class Codehandin {
         changed = true;
     }
 
-    public int getDuedate() {
+    public long getDuedate() {
         return duedate;
     }
